requirements.txt
Pillow~=9.4.0
bitarray~=2.7.3
